+++ src/Exceptions/InvalidFeedItem.php
@@ -7,25 +7,24 @@ use Spatie\Feed\FeedItem;
 
 class InvalidFeedItem extends Exception
 {
-    /** @var \Spatie\Feed\FeedItem|null */
-    public $subject;
+    public ?FeedItem $subject;
 
-    public static function notFeedable($subject)
+    public static function notFeedable($subject): self
     {
         return (new static('Object doesn\'t implement `Spatie\Feed\Feedable`'))->withSubject($subject);
     }
 
-    public static function notAFeedItem($subject)
+    public static function notAFeedItem($subject): self
     {
         return (new static('`toFeedItem` should return an instance of `Spatie\Feed\Feedable`'))->withSubject($subject);
     }
 
-    public static function missingField(FeedItem $subject, $field)
+    public static function missingField(FeedItem $subject, $field): self
     {
         return (new static("Field `{$field}` is required"))->withSubject($subject);
     }
 
-    protected function withSubject($subject)
+    protected function withSubject($subject): self
     {
         $this->subject = $subject;