+++ src/Http/FeedController.php
@@ -2,11 +2,9 @@
 
 namespace Spatie\Feed\Http;
 
-use Illuminate\Support\Arr;
-use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 use Spatie\Feed\Feed;
-use Spatie\Feed\ResolveFeedItems;
+use Spatie\Feed\Helpers\ResolveFeedItems;
 
 class FeedController
 {
@@ -20,27 +18,17 @@ class FeedController
 
         abort_unless($feed, 404);
 
-        $items = $this->resolveFeedItems($feed['items']);
+        $items = ResolveFeedItems::resolve($feed['items']);
 
         return new Feed(
             $feed['title'],
             $items,
             request()->url(),
-            $feed['view'] ?? 'feed::feed',
+            $feed['view'] ?? 'feed::atom',
             $feed['description'] ?? '',
-            $feed['language'] ?? ''
+            $feed['language'] ?? 'en-US',
+            $feed['image'] ?? '',
+            $feed['format'] ?? 'atom',
         );
-    }
-
-    protected function resolveFeedItems($resolver): Collection
-    {
-        $resolver = Arr::wrap($resolver);
-
-        $items = app()->call(
-            array_shift($resolver),
-            $resolver
-        );
-
-        return $items;
     }
 }