+++ tests/DummyItem.php
@@ -8,18 +8,23 @@ use Spatie\Feed\FeedItem;
 
 class DummyItem implements Feedable
 {
+    public function __construct(public int $id = 1)
+    {
+    }
+
     public function toFeedItem(): FeedItem
     {
         return new FeedItem([
-            'id' => 1,
+            'id' => $this->id,
             'title' => 'feedItemTitle',
             'summary' => 'feedItemSummary',
             'enclosure' => 'http://localhost/image1.jpg',
             'enclosureLength' => 31300,
             'enclosureType' => 'image/jpeg',
-            'updated' => Carbon::now(),
-            'link' => 'https://localhost/news/testItem1',
-            'author' => 'feedItemAuthor',
+            'updated' => Carbon::now()->subMinutes($this->id),
+            'link' => 'https://localhost/news/testItem' . $this->id,
+            'authorName' => 'feedItemAuthor',
+            'authorEmail' => 'feedItemAuthor@test.test',
             'category' => 'feedItemCategory',
         ]);
     }