+++ tests/DummyRepository.php
@@ -6,22 +6,25 @@ use Illuminate\Support\Collection;
 
 class DummyRepository
 {
-    /** @var \Illuminate\Support\Collection */
-    public $items;
+    public Collection $items;
 
     public function __construct()
     {
         $this->items = collect([
-            new DummyItem(),
-            new DummyItem(),
-            new DummyItem(),
-            new DummyItem(),
-            new DummyItem(),
+            new DummyItem(1),
+            new DummyItem(2),
+            new DummyItem(3),
+            new DummyItem(4),
+            new DummyItem(5),
         ]);
     }
 
     public function getAll(): Collection
     {
+        if (request()->query->getInt('add', 0) === 1) {
+            $this->items->push(new DummyItem($this->items->count() + 1));
+        }
+
         return $this->items;
     }